okay maybe now! moved auth stuff to rtcclient
